feat(server): 👀 move server files and update paths
